Move "api-boot-plugin-http-convert" to "api-boot"
