+++ api-boot-project/api-boot-autoconfigure/pom.xml
@@ -34,6 +34,8 @@
         <dependency>
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
+            <scope>compile</scope>
+            <optional>true</optional>
         </dependency>
         <!--fastjson-->
         <dependency>
