+++ api-boot-project/api-boot-dependencies/pom.xml
@@ -93,12 +93,6 @@
             <!--ApiBoot Converter-->
             <dependency>
                 <groupId>org.minbox.framework</groupId>
-                <artifactId>api-boot-plugin-http-converter</artifactId>
-                <version>${project.version}</version>
-                <optional>true</optional>
-            </dependency>
-            <dependency>
-                <groupId>org.minbox.framework</groupId>
                 <artifactId>api-boot-starter-http-converter</artifactId>
                 <version>${project.version}</version>
             </dependency>
