--- api-boot-project/api-boot-plugins/api-boot-plugin-http-converter/pom.xml
@@ -1,55 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Copyright [2019] [恒宇少年 - 于起宇]
-  ~
-  ~      Licensed under the Apache License, Version 2.0 (the "License");
-  ~      you may not use this file except in compliance with the License.
-  ~      You may obtain a copy of the License at
-  ~
-  ~          http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~      Unless required by applicable law or agreed to in writing, software
-  ~      distributed under the License is distributed on an "AS IS" BASIS,
-  ~      WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~      See the License for the specific language governing permissions and
-  ~      limitations under the License.
-  ~
-  -->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>api-boot-plugins</artifactId>
-        <groupId>org.minbox.framework</groupId>
-        <version>${revision}</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-    <properties>
-        <main.basedir>${basedir}/../../..</main.basedir>
-    </properties>
-    <artifactId>api-boot-plugin-http-converter</artifactId>
-    <description>
-        ApiBoot Http Converter 插件
-        提供插件的扩展、集成方式
-    </description>
-    <dependencies>
-        <!--ApiBoot Plugin-->
-        <dependency>
-            <groupId>org.minbox.framework</groupId>
-            <artifactId>api-boot-plugin</artifactId>
-        </dependency>
-        <!--fastJson-->
-        <dependency>
-            <groupId>com.alibaba</groupId>
-            <artifactId>fastjson</artifactId>
-            <optional>true</optional>
-        </dependency>
-        <!--Spring Code-->
-        <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-core</artifactId>
-            <optional>true</optional>
-        </dependency>
-    </dependencies>
-</project>
