+++ api-boot-project/api-boot-plugins/pom.xml
@@ -23,7 +23,6 @@
         <module>api-boot-plugin-oauth</module>
         <module>api-boot-plugin-quartz</module>
         <module>api-boot-plugin-resource-load</module>
-        <module>api-boot-plugin-http-converter</module>
         <module>api-boot-plugin-message-push</module>
         <module>api-boot-plugin-rate-limiter</module>
         <module>api-boot-plugin-sequence</module>
