+++ api-boot-project/api-boot-starters/api-boot-starter-http-converter/pom.xml
@@ -32,10 +32,6 @@
             <groupId>org.minbox.framework</groupId>
             <artifactId>api-boot-starter</artifactId>
         </dependency>
-        <dependency>
-            <groupId>org.minbox.framework</groupId>
-            <artifactId>api-boot-plugin-http-converter</artifactId>
-        </dependency>
         <!--fastJson-->
         <dependency>
             <groupId>com.alibaba</groupId>
