+++ api-boot-project/api-boot/src/main/java/org/minbox/framework/api/boot/convert/annotation/ApiBootDecimalAccuracy.java
@@ -15,51 +15,33 @@
  *
  */
 
-package org.minbox.framework.api.boot.plugin.http.converter.filter.annotation;
-
-import org.minbox.framework.api.boot.plugin.http.converter.filter.enums.ValueHidePositionEnum;
+package org.minbox.framework.api.boot.convert.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
+import java.math.BigDecimal;
 
 /**
- * ApiBoot Value Hide Value Filter
+ * ApiBoot Decimal Accuracy
  *
  * @author 恒宇少年
  */
 @Target({ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface ApiBootValueHide {
+public @interface ApiBootDecimalAccuracy {
     /**
-     * hide length
+     * Keep two decimal points
      *
      * @return number
      */
-    int length() default 0;
+    int scale() default 2;
 
     /**
-     * start hide position
+     * BigDecimal round mode
      *
-     * @return start position
+     * @return mode
      */
-    int start() default 0;
-
-    /**
-     * Enumeration of hidden locations
-     * Start hiding after going (exclude value characters of excludeLength)
-     * Start hiding from the back (exclude the value of excludeLength characters)
-     *
-     * @return {@link ValueHidePositionEnum}
-     */
-    ValueHidePositionEnum position() default ValueHidePositionEnum.MIDDLE;
-
-    /**
-     * Replaced hidden placeholders
-     * the default value is "*"
-     *
-     * @return placeholder
-     */
-    String placeholder() default "*";
+    int roundingMode() default BigDecimal.ROUND_DOWN;
 }
