+++ api-boot-samples/api-boot-sample-http-converter/pom.xml
@@ -29,7 +29,7 @@
             <dependency>
                 <groupId>org.minbox.framework</groupId>
                 <artifactId>api-boot-dependencies</artifactId>
-                <version>2.2.2.RELEASE</version>
+                <version>${api-boot.version}</version>
                 <type>pom</type>
                 <scope>import</scope>
             </dependency>
